requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
mysql-connector-python==8.2.0
python-multipart==0.0.6